requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
PyJWT     # simplejwt
sqlparse